feat(Payment): add payment messages
