+++ src/Request/Payments/Command/PaymentAddInterfaceInteractionAction.php
@@ -7,10 +7,19 @@ namespace Commercetools\Core\Request\Payments\Command;
 
 use Commercetools\Core\Model\Common\Context;
 use Commercetools\Core\Request\CustomField\Command\SetCustomTypeAction;
+use Commercetools\Core\Model\CustomField\FieldContainer;
 
 /**
- * @package Commercetools\Core\Request\Payments\Commands
+ * @package Commercetools\Core\Request\Payments\Command
  *
+ * @method string getAction()
+ * @method PaymentAddInterfaceInteractionAction setAction(string $action = null)
+ * @method string getTypeId()
+ * @method PaymentAddInterfaceInteractionAction setTypeId(string $typeId = null)
+ * @method string getTypeKey()
+ * @method PaymentAddInterfaceInteractionAction setTypeKey(string $typeKey = null)
+ * @method FieldContainer getFields()
+ * @method PaymentAddInterfaceInteractionAction setFields(FieldContainer $fields = null)
  */
 class PaymentAddInterfaceInteractionAction extends SetCustomTypeAction
 {
