+++ src/Request/Payments/Command/PaymentSetAmountRefundedAction.php
@@ -10,7 +10,7 @@ use Commercetools\Core\Request\AbstractAction;
 use Commercetools\Core\Model\Common\Money;
 
 /**
- * @package Commercetools\Core\Request\Payments\Commands
+ * @package Commercetools\Core\Request\Payments\Command
  *
  * @method string getAction()
  * @method PaymentSetAmountRefundedAction setAction(string $action = null)
