+++ src/Request/Payments/Command/PaymentSetMethodInfoNameAction.php
@@ -10,7 +10,7 @@ use Commercetools\Core\Request\AbstractAction;
 use Commercetools\Core\Model\Common\LocalizedString;
 
 /**
- * @package Commercetools\Core\Request\Payments\Commands
+ * @package Commercetools\Core\Request\Payments\Command
  *
  * @method string getAction()
  * @method PaymentSetMethodInfoNameAction setAction(string $action = null)
