+++ src/Request/Payments/Command/PaymentTransitionStateAction.php
@@ -9,32 +9,17 @@ use Commercetools\Core\Model\Common\Context;
 use Commercetools\Core\Request\AbstractAction;
 use Commercetools\Core\Model\Customer\CustomerReference;
 use Commercetools\Core\Model\State\StateReference;
+use Commercetools\Core\Request\States\Command\TransitionStateAction;
 
 /**
- * @package Commercetools\Core\Request\Payments\Commands
+ * @package Commercetools\Core\Request\Payments\Command
  *
  * @method string getAction()
  * @method PaymentTransitionStateAction setAction(string $action = null)
  * @method StateReference getState()
  * @method PaymentTransitionStateAction setState(StateReference $state = null)
  */
-class PaymentTransitionStateAction extends AbstractAction
+class PaymentTransitionStateAction extends TransitionStateAction
 {
-    public function fieldDefinitions()
-    {
-        return [
-            'action' => [static::TYPE => 'string'],
-            'state' => [static::TYPE => '\Commercetools\Core\Model\State\StateReference'],
-        ];
-    }
 
-    /**
-     * @param array $data
-     * @param Context|callable $context
-     */
-    public function __construct(array $data = [], $context = null)
-    {
-        parent::__construct($data, $context);
-        $this->setAction('transitionState');
-    }
 }
